% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arco.R
\name{arco}
\alias{arco}
\title{Screen peaks when input data is charcoal areas}
\usage{
arco(
  Seedle.file,
  Smpl.file,
  FireA.file,
  FireC.file,
  n.boot = 10000,
  thresh.prob = 0.95,
  win.width = 1000,
  breakage = FALSE
)
}
\arguments{
\item{Seedle.file}{A data frame with charcoal-particle areas. Should have as
many rows as the number of charcoal particles and two columns in the
following order: Column 1: Top depth of samples;
Column 2: Charcoal-particle areas.}

\item{Smpl.file}{A data frame with as many rows as the number of samples and
seven columns in the following order: \code{CmTop, CmBot, AgeTop, AgeBot,
  volume, charcoal counts, charcoal areas}.}

\item{FireA.file}{Output from peak-detection analysis based on charcoal areas.
Generate it using first the \code{tapas::peak_detection} function (set
argument \code{min_CountP = NULL}), and thereafter the
\code{tapas::tapas_export()} function.}

\item{FireC.file}{Output from peak-detection analysis based on charcoal counts.
Generate it using first the \code{tapas::peak_detection} function (set
argument \code{0 < min_CountP < 1}), and thereafter the
\code{tapas::tapas_export()} function.}

\item{n.boot}{Number of bootstrap samples generated by the function to
obtain a distribution of simulated charcoal-areas. By default,
\code{n.boot = 10000}.}

\item{thresh.prob}{The pth percentile threshold used to separate significant
charcoal-area peaks. By default \code{thresh.prob = 0.95}.}

\item{win.width}{The temporal span of the window from which bootstrap samples
are generated. For each peak to be screened, particles are randomly drawn
(with replacement) from all samples within a focal window, which is
centered on the peak and has a full span of win.width.
By default \code{win.width = 1000}.}

\item{breakage}{Logical. If \code{breakage = FALSE}, plots also
C#/CA-ratios in one of the diagnostic plots. By default
\code{breakage = FALSE}.}
}
\description{
Screens peaks when input data is charcoal-area records. Modified from
the ARCO v1 function, available at url{https://github.com/wfinsinger/ARCO}.
}
\details{
This screening procedure is specific for data sets comprising charcoal
numbers (counts) and areas. It screens the charcoal-area estimates with
respect to the count sums. The method begins with a charcoal-area data set
analysed by existing methods to identify peaks representing fire episodes,
e.g. the \code{peak_detection()} function with the argument
\code{min_CountP = NULL}). To screen these peaks, the method uses bootstrap
resampling of charcoal-particle areas observed in a user-defined subsection
of the data set around each peak (by default \code{win.width = 1000}to
obtain the range of likely charcoal areas for different counts. Peaks with
total area within the likely range of bootstrapped samples (e.g. p > 0.05)
are flagged as potentially unreliable, whereas samples with total area
significantly greater than expected by chance are deemed robust indicators
of past fire episodes.
}
\references{
Finsinger, W., R. Kelly, J. Fevre, and E.K. Magyari. 2014. A guide to
screening charcoal peaks in macrocharcoal-area records for fire episode
reconstructions. The Holocene, 24:1002–1008. doi: 10.1177/0959683614534737

Higuera, P.E., L.B. Brubaker, P.M. Anderson, F.S. Hu, and T.A. Brown. 2009.
Vegetation mediated the impacts of postglacial climate change on fire
regimes in the south-central Brooks Range, Alaska. Ecological Monographs,
79:201–219.
}
\author{
Ryan Kelly

Walter Finsinger
}
