% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/br_sdl.R
\docType{data}
\name{br_sdl}
\alias{br_sdl}
\title{Sedimentary charcoal-particle data (Lake Brazi)}
\format{
A data frame with 843 rows and 2 variables:
\describe{
\item{Depth}{Sample top in centimeters}
\item{SdlArea}{area of a charcoal particle}
}
}
\source{
Finsinger et al. (2009)
\url{https://github.com/wfinsinger/ARCO/blob/master/data_in/arco_Seedle.csv}
}
\usage{
br_sdl
}
\description{
Charcoal-particle data from Lake Brazi, as in:
Finsinger W, Kelly R, Fevre J, Magyari EK (2014) A guide to screening
charcoal peaks in macrocharcoal-area records for fire-episodes
reconstructions. The Holocene 24: 1002–1008: doi:10.1177/0959683614534737.
}
\keyword{datasets}
