% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{co_char_data}
\alias{co_char_data}
\title{TODO: Describe This Data}
\format{
A data frame with 471 rows and 7 variables:
\describe{
\item{cmTop}{Sample top in centimeters}
\item{cmBot}{Sample bottom in centimeters}
\item{AgeTop}{Age of sample Sample top}
\item{AgeBot}{Age of sample Sample bottom}
\item{charvol}{Sample volume in cm³}
\item{char}{Sample count}
\item{dummy_chararea}{dummy variable (charcoal area) to show
that resampling and detrending is done for
all the variables listed in the dataset
(one column for each variable).}
}
}
\source{
TODO: Where does this come from? \url{?}
}
\usage{
co_char_data
}
\description{
This data used for example scripts needs to be described.
}
\keyword{datasets}
