% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CT_GPTpredict.R
\name{CT_GPTpredict}
\alias{CT_GPTpredict}
\title{Cell type annotation for N times with OpenAI GPT models}
\usage{
CT_GPTpredict(N, marker, model = "gpt-4", tissueName = NULL)
}
\arguments{
\item{N}{Repeat run times.}

\item{marker}{A list of genes.}

\item{model}{A valid GPT-4 or GPT-3.5 model name list on https://platform.openai.com/docs/models. Default is 'gpt-4-32k'.}

\item{tissueName}{Optional input of tissue name.}
}
\value{
A data frame of cell types by N times run results.
}
\description{
Annotate cell types by OpenAI GPT models for multiple times with a gene list. One cell type is identified for each element in the list.  Note: system environment should have variable OPENAI_API_KEY = 'your_api_key' or ''. The OpenAI key can be obtained from https://platform.openai.com/account/api-keys. If '', then output the prompt itself. If an actual key is provided, then the output will be the celltype annotations from the GPT model specified by the user.
}
\examples{
data("marker_PBMC")
predict_res<-CT_GPTpredict(N=3,marker= marker_PBMC, tissueName='PBMC', model='gpt-4') 
predict_res
}
\author{
Wen Tang \href{mailto:wxt175@case.edu}{wxt175@case.edu}
}
