% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_overlap_samples.R
\name{get_overlap_depths}
\alias{get_overlap_depths}
\title{Get overlapping sample depths}
\usage{
get_overlap_depths(series)
}
\arguments{
\item{series}{The input data frame. A matrix with the following columns:
\code{CmTop, CmBot, AgeTop, AgeBot, Volume},
and one or more columns with the data
that is being analysed (variables).}
}
\value{
A data frame including only the rows of the overlapping samples
sorted in ascending order by depth. The rownames are inherited from the
input data.frame, and mya thus help finding the flagged samples in raw data
files.
}
\description{
This function makes it easy to see which samples in the input file have
overlapping sample depths. It may be used if the
\code{tapas::check_pretreat()} function flags a warning for the presence of
overlapping sample depths.
}
\details{
The function checks if any \code{CmBot[i] > CmTop[i+1]}. If \code{TRUE},
a data.frame with the overalpping samples is returned.
}
\seealso{
\code{\link[=check_pretreat]{check_pretreat()}}
}
\author{
Walter Finsinger
}
