% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global_Thresh.R
\name{global_thresh}
\alias{global_thresh}
\title{Determine threshold values to extract signal from a detrended timeseries.}
\usage{
global_thresh(
  series = NA,
  proxy = NULL,
  t.lim = NULL,
  thresh.value = 0.95,
  noise.gmm = 1,
  smoothing.yr = NULL,
  keep_consecutive = F,
  minCountP = 0.05,
  MinCountP_window = 150,
  out.dir = NULL,
  plot.global_thresh = T
)
}
\arguments{
\item{series}{The output of the \code{SeriesDetrend()} function.}

\item{proxy}{Set \code{proxy = "VariableName"} to select the variable
for the peak-detection analysis.
If the dataset includes only one variable,
\code{proxy} does not need to be specified.}

\item{t.lim}{Restricted portion of the timeseries.
With \code{t.lim = NULL} (by default),
the analysis is performed with the entire timeseries.}

\item{thresh.value}{Threshold: the nth-percentile of the Gaussian Model
of the noise component.
Defaults to \code{thresh.value = 0.95}.}

\item{noise.gmm}{Specifies which of the two GMM components
should be considered as the noise component.
By default \code{noise.gmm = 1}.}

\item{smoothing.yr}{Width of the moving window
for computing \code{\link{SNI}}.
By default, this value is inherited
from the \code{smoothing.yr} value
set in the \code{SeriesDetrend()} function.}

\item{keep_consecutive}{Logical. When \code{FALSE} (by default),
consecutive peak samples exceeding the threshold
will be removed,
and only the first sample (the oldest) is retained.}

\item{minCountP}{Probability that two resampled counts
could arise from the same Poisson distribution
(defaults to \code{0.05}).
This is used to screen peak samples and remove
any that fails to pass the minimum-count test.
If \code{MinCountP = NULL}, the test is not performed.}

\item{MinCountP_window}{Width of the search window (in years)
used for the minimum-count test.
Defaults to \code{MinCountP_window = 150}.}

\item{out.dir}{Path to the folder where figures will be written to.
Use \code{out.dir = NULL} to plot to current device instead.}

\item{plot.global_thresh}{Logical. If \code{TRUE}, then \code{*.pdf} files
are produced and written to \code{out.dir} folder.}
}
\value{
A list similar to \code{series} with additional data appended.
}
\description{
The script determines threshold values
to decompose a detrended timeseries
into a \emph{noise} and a \emph{signal} component
using a 2-component Gaussian Mixture Model (GMM).
This is based on Phil Higuera's CharThreshLocal.m Matlab code.
It determines a positive and a negative threshold value
for each interpolated sample,
based on the distribution of values within the entire record.
The procedure uses a Gaussian mixture model
with the assumption that the noise component
is normally distributed around 0 (because input values were detrended!).
A figure is generated and saved to the \code{output} directory
and a list is returned with the threshold data for the analyzed proxy.
}
\details{
Requires output from the \code{\link{SeriesDetrend}()} function.
}
