% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Local_Thresh.R
\name{local_thresh}
\alias{local_thresh}
\title{Decompose a detrended timeseries into \emph{noise} and \emph{signal}.}
\usage{
local_thresh(
  series = NA,
  proxy = NULL,
  t.lim = NULL,
  thresh.yr = NULL,
  thresh.value = 0.95,
  smoothing.yr = NULL,
  keep_consecutive = FALSE,
  minCountP = 0.95,
  MinCountP_window = 150,
  out.dir = NULL,
  plot.local_thresh = FALSE
)
}
\arguments{
\item{series}{Output from the \code{\link{SeriesDetrend}()} function.}

\item{proxy}{Set \code{proxy = "VariableName"}
to select the variable for the peak-detection analysis.
If the dataset includes only one variable,
proxy does not need to be specified.}

\item{t.lim}{Allows defining a portion of the time series.
With \code{t.lim = NULL} (by default),
the analysis will be performed using the entire timeseries.}

\item{thresh.yr}{Length of the window width (in years)
from which values are selected
to determine the local threshold.
By default, this value is inherited
from the \code{smoothing.yr} value
set in the \code{\link{SeriesDetrend}()} function.}

\item{thresh.value}{Determines the threshold as the nth-percentile
of the Gaussian Model of the noise component.
Default to \code{thresh.value = 0.95}.}

\item{smoothing.yr}{Width of moving window for computing SNI.}

\item{keep_consecutive}{Logical. If \code{FALSE} (by default),
consecutive peak samples exceeding the threshold
are removed
and only the first (older) sample is retained.}

\item{minCountP}{Probability that two resampled counts could arise
from the same Poisson distribution (default to \code{0.05}).
This is used to screen peak samples
and remove any that fail to pass the minimum-count test.
If \code{MinCountP = NULL}, the test will not be performed.}

\item{MinCountP_window}{Width (in years) of the search window
used for the minimum-count test.
Default to \code{MinCountP_window = 150}.}

\item{out.dir}{Path to the folder where figures are written to.
If \code{out.dir = NULL} (by default),
the plots are emitted to the default device.}

\item{plot.local_thresh}{Logical. If \code{TRUE},
\code{*.pdf} files are produced
and written in the \code{out.dir} folder.
Defaults to \code{FALSE}.}
}
\description{
This is based on Phil Higuera's CharThreshLocal.m Matlab code.
The script determines threshold values
to decompose a detrended timeseries
into a \emph{noise} and a \emph{signal} component
using a 2-component Gaussian Mixture Model (GMM).
It determines a positive and a negative threshold value
for each interpolated sample,
based on the distribution of values
within the selected window (\code{thresh.yr}).
The procedure uses a Gaussian mixture model
on the assumption that the noise component
is normally distributed around 0 (the values were detrended!).
}
\details{
Requires an output from the \code{\link{SeriesDetrend}()} function.
}
