% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{peak_detection}
\alias{peak_detection}
\title{Performs peak-detection analysis of a time series.}
\usage{
peak_detection(
  series = NULL,
  out = "accI",
  proxy = NULL,
  first = NULL,
  last = NULL,
  yrInterp = NULL,
  smoothing_yr = 500,
  detr_type = "mov.median",
  thresh_type = "local",
  thresh_value = 0.95,
  t_lim = NULL,
  noise_gmm = 1,
  keep_consecutive = F,
  min_CountP = 0.05,
  MinCountP_window = 150,
  series_name = NULL,
  out_dir = NULL,
  plotit = TRUE,
  x_lim = t_lim,
  plot_crosses = TRUE,
  plot_x = TRUE,
  plot_neg = FALSE,
  sens = TRUE,
  smoothing_yr_seq = NULL
)
}
\arguments{
\item{series}{A \code{data.frame} with the following columns:
\code{CmTop}, \code{CmBot}, \code{AgeTop},
\code{AgeBot}, \code{Volume},
and one or more columns with the data to be resampled.}

\item{out}{Desired return value: \describe{
\item{"accI"}{the function returns resampled accumulation rates (default)}
\item{"conI"}{the function returns resampled concentrations}
\item{"countI"}{the function returns resampled counts}
}}

\item{proxy}{Select variable for the peak-detection analysis
with \code{proxy = "VariableName"}.
If the dataset includes only one variable,
proxy does not need be specified.}

\item{first, last}{Age boundaries of the resampled time serie.
If unspecified (\code{first=NULL} and \code{last=NULL}),
then resampling is done over the entire sequence,
from \code{min(series$AgeTop)} to max(series$AgeBot).}

\item{yrInterp}{Resolution of the resampled timeseries.}

\item{smoothing_yr}{Smoothing-window width (in years) used for:\itemize{
\item{detrending the data}
\item{smoothing the signal-to-noise index (SNI)}
\item{the window width (in years)
for the local thresholds}
}}

\item{detr_type}{Smoothing function used: \describe{
\item{"rob.loess"}{robust Loess}
\item{"rob.lowess"}{robust Lowess with 4 iterations}
\item{"mov.median"}{moving median (aka. method #4
in Matlab CharAnalysis version)}
}}

\item{thresh_type}{Determines whether to use a single 'global' GMM,
or local GMMs to determine the threshold.
Defaults to \code{thresh_type = "local"}.}

\item{thresh_value}{Determines threshold as the nth-percentile
of the Gaussian Model of the noise component.
Defaults to \code{thresh_value = 0.95}.}

\item{t_lim}{Allows defining a portion of the time series.
With \code{t_lim = NULL} (by default)
the analysis is performed with the entire time series.}

\item{noise_gmm}{Determines which of the two GMM components
should be considered as the noise component.
This is only needed if \code{thresh_type = "global"}.
Defaults to \code{noise_gmm = 1}.}

\item{keep_consecutive}{Logical. If \code{FALSE} (by default),
consecutive peak samples exceeding the threshold
are removed
and only the first (older) sample is retained.}

\item{min_CountP}{Probability that two resampled counts could arise
from the same Poisson distribution.
Defaults to \code{0.95}.
This is used to screen peak samples
and remove any that fail to pass the minimum-count test.
If \code{min_CountP = NULL},
the test will not be performed.}

\item{MinCountP_window}{Width (in years) of the search window
used for the minimum-count test.
Defaults to \code{MinCountP_window = 150}.}

\item{series_name}{A character string defining typically the site name
(\code{NULL} by default).}

\item{out_dir}{Path to the output directory
where \code{*.pdf} figures files are written files.
Defaults to \code{out_dir = NULL},
so the default device is used.}

\item{plotit}{Logical. If \code{TRUE} (by default),
results obtained with the selected settings
are plotted in the display.}

\item{x_lim}{Age limits of the x-axis scale (the time scale).
By default, \code{x_lim = t_lim}.}

\item{plot_crosses}{Logical. If \code{TRUE} (by default),
crosses are added to indicate the location of the events.}

\item{plot_x}{Logical. If \code{FALSE} (by default),
the x-axis labels are omitted.}

\item{plot_neg}{Logical. If \code{FALSE} (by default),
both positive and negative events
are marked with colored shaded areas.}

\item{sens}{Logical. Determines whether a sensitivity analysis is performed.
This involves performing the peak-detection analysis
with different smoothing-window widths (\code{smoothing_yr}),
and plotting boxplots for the SNI distributions,
the Return Intervals, and the number of detected events.}

\item{smoothing_yr_seq}{The smoothing-window widths used
for the sensitivity analysis.
If \code{smoothing_yr_seq = NULL} (by default),
the analysis is performed
for the following smoothing-window widths:
500, 600, 700, 800, 900, 1000 years.}
}
\description{
This wrapper function calls the following functions in correct order:
\code{\link{check_pretreat}()}, \code{\link{pretreatment_data}()},
\code{\link{SeriesDetrend}()}, \code{\link{global_thresh}()},
\code{\link{local_thresh}()}, \code{\link{Plot.Anomalies}()},
and \code{Plot_ReturnIntervals()}.
}
\examples{
\dontrun{
co <- tapas::co_char_data
co_loc <- tapas::peak_detection(co, proxy = "char")
}

}
\author{
Walter Finsinger
}
