% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ar_i.R
\name{plot_ar_i}
\alias{plot_ar_i}
\title{Plot to compare re-sampled and non-resampled accumulation rates}
\usage{
plot_ar_i(
  x,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  frame = TRUE,
  main = NULL,
  ...
)
}
\arguments{
\item{x}{An list returned by the tapas::pretreatment() function.}

\item{xlim}{Optional. Determines the limits of the age scale.
By default (\code{xlim = NULL}), the age-scale limits are in the
form: \code{xlim = c(older limit, younger limit)}.}

\item{ylim}{Optional. Determines the limits of the y-axis scale.
By default, \code{ylim = NULL}.}

\item{xlab}{Optional. A character string specifying the units of the age scale.
By default \code{xlab = NULL} and the x-axis scale label is
written as \code{age}.}

\item{ylab}{Optional. A character string specifying the units of the y-axis.
By default \code{ylab = NULL} and the y-axis scale label is
written as \code{accInit}.}

\item{frame}{Logical. Determines if the plot is framed. By default,
\code{frame = TRUE}.}

\item{main}{Optional. A character string specifying the main plot title.}

\item{\dots}{\dots{}}
}
\description{
Requires an object returned by the tapas::pretreatment() function.
Non-resampled accumulation rates are presented using grey bars. The re-sampled
accumulation rates at equal sampling intervals are presented by a black curve.
Modified from the plot.CHAR() function (paleofire v1.1 R package).
}
\examples{
\dontrun{
# Here, we use the charcoal record from Code Lake (Higuera et al., 2009)
# Load raw charcoal data:
co <- tapas::co_char_data
c <- co[, 6] # charcoal counts
p <- co[, 1:5] # CmTop, CmBot, AgeTop, AgeBot, Volume

# Calculate resampled charcoal accumulation rate (CHAR, as pieces cm-2 yr-1)
co_CHAR <- tapas::pretreatment(params = p, serie = c, Int = TRUE)
tapas::plot_ar_i(co_CHAR)
}
}
\author{
Olivier Blarquez

Walter Finsinger
}
