% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_ReturnIntervals.R
\name{Plot_ReturnIntervals}
\alias{Plot_ReturnIntervals}
\title{Plot return intervals determined based on the Trend And Peaks AnalysiS (tapas).}
\usage{
Plot_ReturnIntervals(series = NULL, x.lim = NULL, plot.x = F, plot.neg = F)
}
\arguments{
\item{series}{The output of \code{\link{global_thresh}()},
\code{\link{local_thresh}()},
or \code{\link{peak_detection}()}.}

\item{x.lim}{Limit age for the x-axis scale (time scale)}

\item{plot.x}{Boolean. If \code{FALSE} (by default),
The x-axis labels are omitted.}

\item{plot.neg}{Boolean. If \code{TRUE} (by default),
the return intervals for both positive and negative anomalies
are plotted.}
}
\description{
Requires output from one of the following functions:
\code{\link{global_thresh}()}, \code{\link{local_thresh}()},
or \code{\link{peak_detection}()}.
}
