% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretreatment.R
\name{pretreatment}
\alias{pretreatment}
\title{Calculate re-sampled particles' accumulation rates for sediment records}
\usage{
pretreatment(
  params,
  serie,
  Int = TRUE,
  first = NULL,
  last = NULL,
  yrInterp = NULL
)
}
\arguments{
\item{params}{A matrix with the following columns: CmTop, CmBot, AgeTop,
AgeBot, Volume, in the same order.}

\item{serie}{Numerical. A record to be transformed in accumulation rates.
Could be the number, the area, or the volume of particles in samples.}

\item{Int}{Logical. Specifies whether the function interpolates missing
values. By default \code{Int = TRUE}. Missing values are identified when
the value of \code{serie} is equal to -999 or NA, or when the sample
Volume = 0.}

\item{first, last}{Date of the first, last sample for accumulation rate
calculation, if NULL first, last are automatically specified as the the
minimum and maximum ages of the record respectively.}

\item{yrInterp}{Temporal resolution of the interpolated accumulation rates,
if NULL, yrInterp is automatically specified as the median resolution of the
record.}
}
\value{
Returns an object of the class "proxy_ar_i" having the output structure:
\item{cmI}{interpolated depths}
\item{ybpI}{interpolated ages}
\item{accI}{accumulation rates}
}
\description{
This is the R version of the CharPretreatment.m function, which was
originally developed by Philip Higuera for "CharAnalysis" (see
https://github.com/phiguera/CharAnalysis/tree/master). It has been translated
from the Matlab-coded CharPretreatment.m function by Olivier Blarquez and was
included in the paleofire v1.2.4 R package (https://github.com/oblarquez/paleofire).
The function is now included in the tapas R package because the paleofire R
package is currently not being updated.
}
\examples{
\dontrun{
# Load raw charcoal data:
co <- tapas::co_char_data
c <- co[, 6] # charcoal counts
p <- co[, 1:5] # CmTop, CmBot, AgeTop, AgeBot, Volume

# Calculates charcoal accumulation rate (CHAR, as pieces cm-2 yr-1)
co_CHAR <- tapas::pretreatment(params = p, serie = c, Int = TRUE)
plot(co_CHAR)
}
}
\author{
Olivier Blarquez
}
