% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_imagej_results.R
\name{read_imagej_results}
\alias{read_imagej_results}
\title{Reads results of charcoal analyses made with ImageJ}
\usage{
read_imagej_results(
  folder_in = NULL,
  metadata = c("site_id", "depth", "image_id"),
  image_id_default = "a"
)
}
\arguments{
\item{folder_in}{Sets the path to the directory where the *.csv files are
located (e.g. "./data_in/")}

\item{metadata}{A vector that describes the metadata stored in the file names.
By default, metadata = c("core_id", "depth", "image_id"),
for file names formatted as "CoreID_Depth_ImageID.csv".
However, other file name formats are possible. See the
Description here below.}

\item{image_id_default}{Specifies the value or character that shall be
added if the file name does not include the
information for the ImageID. By default,
\code{image_id_default = "a"}.}
}
\value{
A data.frame with the specified metadata columns, and all columns that
are common to all input files.
}
\description{
The function loads spreadsheets obtained with ImageJ (https://imagej.net/)
that were stored as *.csv files (comma-separated) in a common folder whose
path is specified by the \code{folder_in} argument.
The function expects that the file names include a set of information that
identify the samples unequivocally, and that the information is given as
strings separated by a dash (-) or an underscore (_) (do not use empty
spaces (space bar) to separate the metadata).

By default the following information is expected:
\itemize{
\item a site identifier;
\item a sample identifier, for instance the master core depth;
\item an image identifier (a character string) in case several images were
taken for a sample. If only one image was taken from a sample, the
image identifier is optional (an "a" will be added if the file name does
not include this information).
}

However, the function allows to use a different set and number of
identifiers. Possible alternatives may include:
\itemize{
\item "CoreID_SectionID_Depth_ImageID.csv" or
\item "CoreID_SectionID_DepthInDrive_ImageID.csv".
}

If the file names include a "depth" value, make sure to take note as to
which depth values are mentioned (top depth, mid depth, bottom depth), as
well as to which depth scale the values are referring to (e.g. the master
core depth, depth in drive, or other).
Further, make sure the file names include useful information to link
the data with other data types from the same sediment archive,
such as the chronology or other proxy data.
}
\author{
Walter Finsinger
}
