% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeriesDetrend.R
\name{SeriesDetrend}
\alias{SeriesDetrend}
\title{Detrend time series.}
\usage{
SeriesDetrend(
  series = NULL,
  smoothing.yr = NULL,
  detr.type = "rob.loess",
  out.dir = NULL,
  plot_pdf = T
)
}
\arguments{
\item{series}{Output of pretreatment_data().}

\item{smoothing.yr}{Smoothing window width.}

\item{detr.type}{Smoothing function used: \describe{
\item{"rob.loess"}{robust Loess}
\item{"rob.lowess"}{robust Lowess with 4 iterations}
\item{"mov.mode"}{moving median (aka. method #4
in Matlab CharAnalysis version)}
}}

\item{out.dir}{Folder to output the generated figures to.
Defaults to NULL, and the figures are generated
to default device instead.}

\item{plot_pdf}{Logical. If \code{TRUE}, a \code{*.pdf} file
will be written in the \code{out.dir} folder.}
}
\value{
A list similar to \code{series} with additional detrended data
appended.
}
\description{
Processes the output of pretreatment_data() function
to de-trend the time series.
}
