% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNI.R
\name{SNI}
\alias{SNI}
\title{Calculate signal-to-noise index (SNI) for paleoecological timeseries records.}
\usage{
SNI(ProxyData, BandWidth)
}
\arguments{
\item{ProxyData}{Matrix of input data with one row per sample,
containing: \describe{
\item{Column 1}{age associated with the sample (yr)}
\item{Column 2}{proxy accumulation rate (e.g. CHAR)
of the sample (pieces/cm^2/yr)}
\item{Column 3}{threshold value (pieces/cm^2/yr)}
}}

\item{BandWidth}{Width of moving window for computing SNI.}
}
\value{
The function was modified by Walter Finsinger
to return only a subset
of the original \code{SNI_output} list: \describe{
\item{\code{$SNI_sm}}{Lowess-smoothed SNI values.}
\item{\code{$SNI_raw}}{Raw SNI values.}
}
In the original script,
the returned data list contained the computed SNI
and related data,
with one row for each row in the input variable (ProxyData):
\describe{
\item{\code{$SNI_sm}}{The smoothed SNI computed for each sample.}
\item{\code{$winInd}}{Indexes of the first and last samples
included in each moving window.
E.g. \code{SNI_output$winInd(X) == [A, B]}
indicates that the moving window used
to calculate SNI for the \code{X}th sample
contained all samples between \code{A} and \code{B},
inclusive.}
\item{\code{$popN}}{The CHAR values
of all samples in the noise (N) population
(samples in the moving window
with CHAR below threshold)}
\item{\code{$popS}}{The CHAR values
of all samples in the signal (S) population
(samples in the moving window
with CHAR at or above threshold)}
\item{\code{$meanN}}{Mean CHAR of the samples in popN}
\item{\code{$stdN}}{standard deviation of the samples in popN}
\item{\code{$CF}}{The \emph{correction factor}
used in computing SNI.
Equal to (v - 2)/v,
where v is the number of samples in popN}
}
}
\description{
This function computes SNI as described in Kelly et al. 2011.
Note that your data must be interpolated
to constant sample resolution (yr/sample)
before input to the function.
The function makes no assumption about prior analysis on the input series,
i.e. any background and threshold methods may be used.
However, input data should still be consistent
with the interpretation that a value (column 2)
greater than the corresponding threshold value (column 3)
is a \emph{signal} sample,
whereas a value below the threshold is \emph{noise}.
Refer to Kelly et al. 2010 for details and discussion.
}
\references{
Supplementary materials to the publication:
Kelly, R., Higuera, P., Barrett, C., & Hu, F. (2011).
A signal-to-noise index to quantify
the potential for peak detection in sediment–charcoal records.
Quaternary Research, 75(1), 11-17.
\url{https://doi.org/10.1016/j.yqres.2010.07.011}
}
\seealso{
Supplementary materials:
\url{https://static.cambridge.org/content/id/urn:cambridge.org:id:article:S0033589400006785/resource/name/S0033589400006785sup001.txt}
}
\author{
Ryan Kelly (University of Illinois, USA)
}
