% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapas2mgcv.R
\name{tapas2mgcv}
\alias{tapas2mgcv}
\title{Extract data from a pretreatment_data() list to determine trend with a GAM.}
\usage{
tapas2mgcv(series = NULL)
}
\arguments{
\item{series}{A list that was generated with the
\code{tapas::pretreatment_data()} function.}
}
\value{
The returned data frame includes the following columns: cm_top, cm_bot,
age_top, age_bot, volI, and in addition one column for each of the variables
that were included in the input data frame for the
\code{tapas::pretreatment_data()} function.
The returned age corresponds to the age_top of the samples.
}
\description{
Gathers data from a list produced with the \code{tapas::pretreatment_data()}
function, and returns a data frame that can be used to determine the trend
based on a Generalized Additive Model with the \code{mgcv::gam()} function.
}
\details{
The \code{tapas::pretreatment_data()} function sometimes may generate NA
values in the cm_top and cm_bot columns for the first and/or the last samples
in a temporal series. The \code{tapas::tapasa2mgcv()} function removes these
NA values automatically by taking the sediment-accumulation rate of adjacent
binned samples.
}
\seealso{
\code{\link[=mgcv2tapas]{mgcv2tapas()}}
}
\author{
Walter Finsinger
}
