% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_tapas.R
\name{tapas_export}
\alias{tapas_export}
\title{Prepares an output data frame to summarize the results of the peak-detection analysis.}
\usage{
tapas_export(series = NULL)
}
\arguments{
\item{series}{The output of \code{\link{global_thresh}()},
\code{\link{local_thresh}()},
or \code{\link{peak_detection}()}.}
}
\value{
A data frame.
}
\description{
Requires output from one of the following functions:
\code{\link{global_thresh}()}, \code{\link{local_thresh}()},
or \code{\link{peak_detection}()}.
}
