% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tapas_res.R
\name{tapas_pipeline}
\alias{tapas_pipeline}
\title{Run the Full Analysis Pipeline for a Specific Tissue}
\usage{
tapas_pipeline(
  cell_type_lists,
  model = "gpt-4",
  N = 20,
  tissue = c("PBMC", "Heart", "Lung", "Kidney", "Brain", "Breast", "Liver")
)
}
\arguments{
\item{cell_type_lists}{A list or matrix of predicted cell type lists generated by the CT_GPTpredict.}

\item{model}{The GPT model to use (default is 'gpt-4').}

\item{N}{An integer for the repeat run times.}

\item{tissue}{A character string specifying the tissue type ("PBMC", "Heart", "Lung", or "Kidney").}
}
\value{
A list containing the p-values.
}
\description{
This function assigns cell types using an OpenAI GPT model, computes weighted average distances,
and calculates p-values based on a gamma distribution from a bootstrap. The appropriate tree, distance matrix,
and gamma parameters (shape and rate) are selected automatically based on the tissue type.
}
\examples{
data(list = c("heart_df","heart_distance_mtx","kidney_df","kidney_distance_mtx","lung_df","lung_distance_mtx","pbmc_df","pbmc_distance_mtx",
"brain_df","brain_distance_mtx_na","breast_df","breast_distance_mtx_na","liver_df","liver_distance_mtx_na","marker_PBMC"),package = "Tapas")
predict_res<-CT_predict(N=3,marker= marker_PBMC, tissueName='PBMC', model='gpt-4',provider='gpt') 
p_res<-tapas_pipeline(predict_res,N=3,tissue="PBMC")
p_res
}
