% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pretreatment.R
\name{check_pretreat}
\alias{check_pretreat}
\title{Check data frame used for \code{paleofire::pretreatment()} function.}
\usage{
check_pretreat(series)
}
\arguments{
\item{series}{The input data frame. A matrix with the following columns:
\code{CmTop, CmBot, AgeTop, AgeBot, Volume},
and one or more columns with the data
which should be resampled (variables).}
}
\description{
This functions does some check-up on the data frame used for
the \code{paleofire::pretreatment()} function,
which requires that the depth and age scales are continuous.
In other words, for every i-th row, it requires that \itemize{
\item{\code{CmBot[i] > CmTop[i]}}
\item{\code{AgeBot[i] > AgeTop[i]}}
\item{there shouldn't be duplicate values
in the \code{CmTop, CmBot, AgeTop}, and \code{AgeBot} columns}
\item{\code{CmBot[i] == CmTop[i+1]}}
\item{\code{AgeBot[i] == AgeTop[i+1]}}
}
}
\details{
If any of the following is true: \itemize{
\item{any \code{CmBot[i] < CmTop[i]}}
\item{\code{AgeBot[i] < AgeTop[i]}}
\item{there are duplicate values in one of the columns \code{CmTop, CmBot}}
}
.. then the function returns a fatal error and stops.

The function also fixes a few things:

If any \code{CmBot[i] < CmTop[i+1]}, then
the function adds a new row \code{[j = i+1]}
to fill in the missing \code{CmTop} and \code{CmBot} values.
Thus, the added samples will have \code{CmTop[j] == CmBot[i]}
and \code{CmBot[j] == CmTop[i+1]}.
The age scale values will be:
\code{AgeTop[j] == AgeBot[i]}
and \code{AgeBot[j] == AgeTop[i+1]}.

If after these checks any \code{AgeTop[i] == AgeBot[i]}, then
the function \emph{removes the flagged rows},
and \emph{creates new} corrected \code{CmTop} and \code{CmBot} scales
that exclude slumps,
such that \code{CmBot[i] > CmTop[i]}
and \code{CmBot[i] == CmTop[i+1]}.
}
