% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{co_char_data}
\alias{co_char_data}
\title{Lake-sediment charcoal data from Code lake, as in: Higuera, P. E., L. B. Brubaker, P. M. Anderson, F. S. Hu, and T. A. Brown. 2009. Vegetation mediated the impacts of postglacial climate change on fire regimes in the south-central Brooks Range, Alaska. Ecological Monographs 79:201-219.}
\format{
A data frame with 471 rows and 7 variables:
\describe{
\item{cmTop}{Sample top in centimeters}
\item{cmBot}{Sample bottom in centimeters}
\item{AgeTop}{Age of sample Sample top}
\item{AgeBot}{Age of sample Sample bottom}
\item{charvol}{Sample volume in cm³}
\item{char}{Sample count}
\item{dummy_chararea}{dummy variable (charcoal area) to show
that resampling and detrending is done for
all the variables listed in the dataset
(one column for each variable).}
}
}
\source{
Higuera et al. (2009) \url{https://figshare.com/articles/dataset/Higuera_et_al_2009_lake_sediment_pollen_and_charcoal_data/984310/4}
}
\usage{
co_char_data
}
\description{
Lake-sediment charcoal data from Code lake, as in: Higuera, P. E., L. B. Brubaker, P. M. Anderson, F. S. Hu, and T. A. Brown. 2009. Vegetation mediated the impacts of postglacial climate change on fire regimes in the south-central Brooks Range, Alaska. Ecological Monographs 79:201-219.
}
\keyword{datasets}
