% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretreatment_data.R
\name{pretreatment_data}
\alias{pretreatment_data}
\title{Pre-process data by resampling all columns.}
\usage{
pretreatment_data(
  series = NULL,
  out = "accI",
  series.name = NA,
  first = NULL,
  last = NULL,
  yrInterp = NULL
)
}
\arguments{
\item{series}{A matrix with the following first columns:
\code{c("CmTop", "CmBot", "AgeTop", "AgeBot", "Volume")}
and additional columns with the data which should be resampled.}

\item{out}{Desired return value: \describe{
\item{"accI"}{the function returns resampled accumulation rates}
\item{"conI"}{the function returns resampled concentrations}
\item{"countI"}{the function returns resampled counts}
}}

\item{series.name}{A string defining typically the site name. Defaults to
\code{NA}.}

\item{first, last}{Age boundaries of the resampled time serie.
If unspecified (\code{first=NULL} and \code{last=NULL}),
then resampling is done over the entire sequence,
from \code{min(series$AgeTop)} to max(series$AgeBot).}

\item{yrInterp}{Resolution of the resampled timeseries.}
}
\value{
A list with the resampled (binned) data according to
\code{out} parameter.
}
\description{
Uses the paleofire::pretreatment() function
to virtually resample timeseries at equal sampling intervals.
}
